requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Flask-Migrate
Flask-Cors
